+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Fixes
 
-- Add the compatibility with an entry name with slashes and the `outputPath` option [#81](https://github.com/yoriiis/chunks-webpack-plugin/pull/81)
+- Fix incompatibility with `outputPath` and slashes on entry name #81 [#81](https://github.com/yoriiis/chunks-webpack-plugin/pull/81)
 
 ## 7.0.2